+++ servers/all/$server.ts
@@ -74,23 +74,17 @@ export type ServerMethods<T extends AspidaMethods, U extends Record<string, any>
   ) => ServerResponse<T[K]> | Promise<ServerResponse<T[K]>>
 }
 
-const parseNumberTypeQueryParams = (numberTypeParamsFn: (query: any) => ([string, boolean, boolean][])): preValidationHookHandler => (req, reply, done) => {
+const parseNumberTypeQueryParams = (numberTypeParams: [string, boolean, boolean][]): preValidationHookHandler => (req, reply, done) => {
   const query: any = req.query
-  const numberTypeParams = numberTypeParamsFn(query)
 
   for (const [key, isOptional, isArray] of numberTypeParams) {
-    const param = query[key]
+    const param = isArray ? (query[`${key}[]`] ?? query[key]) : query[key]
 
     if (isArray) {
       if (!isOptional && param === undefined) {
         query[key] = []
       } else if (!isOptional || param !== undefined) {
-        if (!Array.isArray(param)) {
-          reply.code(400).send()
-          return
-        }
-
-        const vals = (param as string[]).map(Number)
+        const vals = (Array.isArray(param) ? param : [param]).map(Number)
 
         if (vals.some(isNaN)) {
           reply.code(400).send()
@@ -99,6 +93,8 @@ const parseNumberTypeQueryParams = (numberTypeParamsFn: (query: any) => ([string
 
         query[key] = vals as any
       }
+
+      delete query[`${key}[]`]
     } else if (!isOptional || param !== undefined) {
       const val = Number(param)
 
@@ -111,6 +107,50 @@ const parseNumberTypeQueryParams = (numberTypeParamsFn: (query: any) => ([string
     }
   }
 
+  done()
+}
+
+const parseBooleanTypeQueryParams = (booleanTypeParams: [string, boolean, boolean][]): preValidationHookHandler => (req, reply, done) => {
+  const query: any = req.query
+
+  for (const [key, isOptional, isArray] of booleanTypeParams) {
+    const param = isArray ? (query[`${key}[]`] ?? query[key]) : query[key]
+
+    if (isArray) {
+      if (!isOptional && param === undefined) {
+        query[key] = []
+      } else if (!isOptional || param !== undefined) {
+        const vals = (Array.isArray(param) ? param : [param]).map(p => p === 'true' ? true : p === 'false' ? false : null)
+
+        if (vals.some(v => v === null)) {
+          reply.code(400).send()
+          return
+        }
+
+        query[key] = vals as any
+      }
+
+      delete query[`${key}[]`]
+    } else if (!isOptional || param !== undefined) {
+      const val = param === 'true' ? true : param === 'false' ? false : null
+
+      if (val === null) {
+        reply.code(400).send()
+        return
+      }
+
+      query[key] = val as any
+    }
+  }
+
+  done()
+}
+
+const callParserIfExistsQuery = (parser: preValidationHookHandler): preValidationHookHandler => (req, reply, done) =>
+  Object.keys(req.query as any).length ? parser(req, reply, done) : done()
+
+const normalizeQuery: preValidationHookHandler = (req, _, done) => {
+  req.query = JSON.parse(JSON.stringify(req.query))
   done()
 }
 
@@ -209,7 +249,9 @@ export default (fastify: FastifyInstance, options: FrourioOptions = {}) => {
       onRequest: [...hooks0.onRequest, ctrlHooks0.onRequest],
       preParsing: hooks0.preParsing,
       preValidation: [
-        parseNumberTypeQueryParams(query => !Object.keys(query).length ? [] : [['requiredNum', false, false], ['optionalNum', true, false], ['optionalNumArr', true, true], ['emptyNum', true, false], ['requiredNumArr', false, true]]),
+        callParserIfExistsQuery(parseNumberTypeQueryParams([['requiredNum', false, false], ['optionalNum', true, false], ['optionalNumArr', true, true], ['emptyNum', true, false], ['requiredNumArr', false, true]])),
+        callParserIfExistsQuery(parseBooleanTypeQueryParams([['bool', false, false], ['optionalBool', true, false], ['boolArray', false, true], ['optionalBoolArray', true, true]])),
+        normalizeQuery,
         createValidateHandler(req => [
           Object.keys(req.query as any).length ? validateOrReject(Object.assign(new Validators.Query(), req.query as any), validatorOptions) : null
         ])
@@ -224,8 +266,10 @@ export default (fastify: FastifyInstance, options: FrourioOptions = {}) => {
       onRequest: [...hooks0.onRequest, ctrlHooks0.onRequest],
       preParsing: hooks0.preParsing,
       preValidation: [
-        parseNumberTypeQueryParams(() => [['requiredNum', false, false], ['optionalNum', true, false], ['optionalNumArr', true, true], ['emptyNum', true, false], ['requiredNumArr', false, true]]),
+        parseNumberTypeQueryParams([['requiredNum', false, false], ['optionalNum', true, false], ['optionalNumArr', true, true], ['emptyNum', true, false], ['requiredNumArr', false, true]]),
+        parseBooleanTypeQueryParams([['bool', false, false], ['optionalBool', true, false], ['boolArray', false, true], ['optionalBoolArray', true, true]]),
         formatMultipartData([]),
+        normalizeQuery,
         createValidateHandler(req => [
           validateOrReject(Object.assign(new Validators.Query(), req.query as any), validatorOptions),
           validateOrReject(Object.assign(new Validators.Body(), req.body as any), validatorOptions)
@@ -273,7 +317,7 @@ export default (fastify: FastifyInstance, options: FrourioOptions = {}) => {
     {
       onRequest: hooks0.onRequest,
       preParsing: hooks0.preParsing,
-      preValidation: parseNumberTypeQueryParams(query => !Object.keys(query).length ? [] : [['limit', true, false]])
+      preValidation: callParserIfExistsQuery(parseNumberTypeQueryParams([['limit', true, false]]))
     },
     methodToHandler(controller4.get)
   )