+++ src/createControllersText.ts
@@ -362,32 +362,29 @@ export default (appDir: string, project: string) => {
 
         const genResSchemaText = (method: LowerHttpMethod) =>
           `schema: { response: responseSchema${controllers.filter(c => c[2]).length}.${method} }`
+        const getSomeTypeQueryParams = (typeName: string, query: ts.Symbol) =>
+          checker
+            .getTypeOfSymbolAtLocation(query, query.valueDeclaration)
+            .getProperties()
+            .map(p => {
+              const typeString = checker.typeToString(
+                checker.getTypeOfSymbolAtLocation(p, p.valueDeclaration)
+              )
+              return typeString === typeName || typeString === `${typeName}[]`
+                ? `['${p.name}', ${p.declarations.some(d =>
+                    d.getChildren().some(c => c.kind === ts.SyntaxKind.QuestionToken)
+                  )}, ${typeString === `${typeName}[]`}]`
+                : null
+            })
+            .filter(Boolean)
 
         results.push(
           methods
             .map(m => {
               const props = checker.getTypeOfSymbolAtLocation(m, m.valueDeclaration).getProperties()
               const query = props.find(p => p.name === 'query')
-              const numberTypeQueryParams =
-                query &&
-                checker
-                  .getTypeOfSymbolAtLocation(query, query.valueDeclaration)
-                  .getProperties()
-                  .map(p => {
-                    const typeString = checker.typeToString(
-                      checker.getTypeOfSymbolAtLocation(p, p.valueDeclaration)
-                    )
-                    return typeString === 'number'
-                      ? `['${p.name}', ${p.declarations.some(d =>
-                          d.getChildren().some(c => c.kind === ts.SyntaxKind.QuestionToken)
-                        )}, false]`
-                      : typeString === 'number[]'
-                      ? `['${p.name}', ${p.declarations.some(d =>
-                          d.getChildren().some(c => c.kind === ts.SyntaxKind.QuestionToken)
-                        )}, true]`
-                      : null
-                  })
-                  .filter(Boolean)
+              const numberTypeQueryParams = query && getSomeTypeQueryParams('number', query)
+              const booleanTypeQueryParams = query && getSomeTypeQueryParams('boolean', query)
               const validateInfo = [
                 { name: 'query', val: query },
                 { name: 'body', val: props.find(p => p.name === 'reqBody') },
@@ -419,14 +416,23 @@ export default (appDir: string, project: string) => {
                 .map(key => {
                   if (key === 'preValidation') {
                     const texts = [
-                      numberTypeQueryParams && numberTypeQueryParams.length
-                        ? `parseNumberTypeQueryParams(${
-                            query?.declarations.some(
-                              d => d.getChildAt(1).kind === ts.SyntaxKind.QuestionToken
-                            )
-                              ? 'query => !Object.keys(query).length ? [] :'
-                              : '() =>'
-                          } [${numberTypeQueryParams.join(', ')}])`
+                      numberTypeQueryParams?.length
+                        ? query?.declarations.some(
+                            d => d.getChildAt(1).kind === ts.SyntaxKind.QuestionToken
+                          )
+                          ? `callParserIfExistsQuery(parseNumberTypeQueryParams([${numberTypeQueryParams.join(
+                              ', '
+                            )}]))`
+                          : `parseNumberTypeQueryParams([${numberTypeQueryParams.join(', ')}])`
+                        : '',
+                      booleanTypeQueryParams?.length
+                        ? query?.declarations.some(
+                            d => d.getChildAt(1).kind === ts.SyntaxKind.QuestionToken
+                          )
+                          ? `callParserIfExistsQuery(parseBooleanTypeQueryParams([${booleanTypeQueryParams.join(
+                              ', '
+                            )}]))`
+                          : `parseBooleanTypeQueryParams([${booleanTypeQueryParams.join(', ')}])`
                         : '',
                       isFormData && reqBody
                         ? `formatMultipartData([${checker
@@ -451,6 +457,14 @@ export default (appDir: string, project: string) => {
                             .join(', ')}])`
                         : '',
                       ...genHookTexts('preValidation'),
+                      ...(query &&
+                      [
+                        ...(numberTypeQueryParams ?? []),
+                        ...(booleanTypeQueryParams ?? [])
+                      ].some(t => t?.endsWith('true]')) &&
+                      validateInfo.length
+                        ? ['normalizeQuery']
+                        : []),
                       validateInfo.length
                         ? `createValidateHandler(req => [
 ${validateInfo
@@ -476,7 +490,7 @@ ${validateInfo
                     return texts.length
                       ? `${key}: ${
                           texts.length === 1
-                            ? texts[0].replace('...', '')
+                            ? texts[0].replace(/^\.+/, '')
                             : `[\n        ${texts.join(',\n        ')}\n      ]`
                         }`
                       : ''